Developement

+++ server/db/migrations/20230405222114_attendees.js
@@ -1,19 +1,11 @@
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
 exports.up = function (knex) {
   return knex.schema.createTable('attendees', (table) => {
     table.increments('id').primary()
     table.string('name')
-    table.integer('wage')
+    table.decimal('wage')
   })
 }
 
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
 exports.down = function (knex) {
   return knex.schema.dropTable('attendees')
 }
